enable warnings and get rid of some

+++ tiled_image.h
@@ -9,7 +9,7 @@
 namespace {
 struct TiledImage {
     static constexpr int image_size = 32 * 1000L + 34;
-    static constexpr uint16_t image_wdwidth = 320 * 4 / sizeof(uint8_t);
+    static constexpr int image_wdwidth = 320 * 4 / sizeof(uint8_t);
     static constexpr uint16_t width = 40;
     static constexpr uint16_t height = 40;
